# Use lock-free random number generator for node height selection